+++ minecraft_model_reader/__init__.py
@@ -1,7 +1,6 @@
 import os
 from minecraft_model_reader.api.log import log
 from minecraft_model_reader.api.mesh.block.block_mesh import BlockMesh
-from minecraft_model_reader.api.resource_pack import BaseRP, BaseRPHandler
-from minecraft_model_reader.java.java_rp_handler import JavaRP, JavaRPHandler
+from minecraft_model_reader.api.resource_pack import BaseResourcePack, BaseResourcePackManager
 
 path = os.path.dirname(__file__)
