+++ minecraft_model_reader/api/resource_pack/base/__init__.py
@@ -1,0 +1,2 @@
+from .resource_pack import BaseResourcePack
+from .resource_pack_manager import BaseResourcePackManager
