+++ minecraft_model_reader/api/resource_pack/base/resource_pack.py
