+++ minecraft_model_reader/api/resource_pack/java/java_block_model.py
